Move vulkan_gui_util to iree/testing/vulkan
