+++ iree/samples/vulkan/BUILD
@@ -47,21 +47,13 @@ cc_binary(
     ],
     deps = [
         ":simple_mul_bytecode_module_cc",
-        "//iree/base:api",
-        "//iree/base:logging",
         "//iree/base:main",
-        "//iree/hal:api",
-        "//iree/hal/vulkan:api",
         "//iree/modules/hal",
+        "//iree/testing/vulkan:vulkan_gui_util",
         "//iree/vm",
         "//iree/vm:bytecode_module",
         "//iree/vm:ref_cc",
         "@com_google_absl//absl/base:core_headers",
         "@com_google_absl//absl/types:span",
-        "@dear_imgui",
-        "@dear_imgui//:imgui_sdl_vulkan",
-        "@iree_vulkan_headers//:vulkan_headers_no_prototypes",
-        "@sdl2//:SDL2",
-        "@vulkan_sdk//:sdk",
     ],
 )
